chore: rename some components
