+++ packages/desktop/views/dashboard/governance/views/ProposalDetailsView.svelte
